agent_core/requirements.txt
numpy==2.1.0
pymavlink==2.4.38
pyserial==3.5
PyYAML==6.0.2
pyzmq==23.2.0
websocket-client
wsaccel